refactor(backend): :recycle: refatoração da rota
refatoração da rota uploadExtrato para a estutura de arquivos do modelo msc
